Added new abilities Changed Divine Shield to Magic attack Added speed increase and dex increase Altered combat slider script

+++ MajorProject/Assets/Scripts/MagicAttack.cs
@@ -51,8 +51,8 @@ public class MagicAttack : WeaponBase {
     {
         DoWithFuncs funcToDo = dowithFunc[(int)m_doWith];
         if (m_magicAlignment == Aliginment.Law)
-            return funcToDo((int)law, alignmentRange);
+            return funcToDo(law, alignmentRange);
         else
-            return funcToDo((int)light, alignmentRange);
+            return funcToDo(light, alignmentRange);
     }
 }
